+++ packages/adapter-node/rollup.config.js
@@ -4,31 +4,15 @@ import json from '@rollup/plugin-json';
 
 export default [
 	{
-		input: 'src/middlewares.js',
-		output: {
-			file: 'files/middlewares.js',
-			format: 'esm',
-			sourcemap: true
+		input: {
+			index: 'src/index.js',
+			handler: 'src/handler.js'
 		},
-		plugins: [nodeResolve(), commonjs(), json()],
-		external: ['../output/server/app.js', ...require('module').builtinModules]
-	},
-	{
-		input: 'src/index.js',
 		output: {
-			file: 'files/index.js',
-			format: 'esm',
-			sourcemap: true
-		},
-		plugins: [nodeResolve(), commonjs(), json()],
-		external: ['./middlewares.js', './env.js', ...require('module').builtinModules]
-	},
-	{
-		input: 'src/shims.js',
-		output: {
-			file: 'files/shims.js',
+			dir: 'files',
 			format: 'esm'
 		},
-		external: ['module']
+		plugins: [nodeResolve(), commonjs(), json()],
+		external: ['APP', 'MANIFEST', ...require('module').builtinModules]
 	}
 ];