+++ packages/kit/src/core/config/index.spec.js
@@ -29,19 +29,25 @@ test('fills in defaults', () => {
 				template: 'src/app.html'
 			},
 			floc: false,
+			headers: {
+				host: null,
+				protocol: null
+			},
 			host: null,
-			hostHeader: null,
 			hydrate: true,
 			package: {
 				dir: 'package',
 				emitTypes: true
 			},
-			serviceWorker: {},
+			serviceWorker: {
+				register: true
+			},
 			paths: {
 				base: '',
 				assets: ''
 			},
 			prerender: {
+				concurrency: 1,
 				crawl: true,
 				enabled: true,
 				entries: ['*'],
@@ -49,6 +55,7 @@ test('fills in defaults', () => {
 				onError: 'fail',
 				pages: undefined
 			},
+			protocol: null,
 			router: true,
 			ssr: true,
 			target: null,
@@ -129,19 +136,25 @@ test('fills in partial blanks', () => {
 				template: 'src/app.html'
 			},
 			floc: false,
+			headers: {
+				host: null,
+				protocol: null
+			},
 			host: null,
-			hostHeader: null,
 			hydrate: true,
 			package: {
 				dir: 'package',
 				emitTypes: true
 			},
-			serviceWorker: {},
+			serviceWorker: {
+				register: true
+			},
 			paths: {
 				base: '',
 				assets: ''
 			},
 			prerender: {
+				concurrency: 1,
 				crawl: true,
 				enabled: true,
 				entries: ['*'],
@@ -149,6 +162,7 @@ test('fills in partial blanks', () => {
 				onError: 'fail',
 				pages: undefined
 			},
+			protocol: null,
 			router: true,
 			ssr: true,
 			target: null,