+++ packages/kit/src/runtime/app/stores.js
@@ -56,7 +56,7 @@ export const navigating = {
 };
 
 /** @param {string} verb */
-const error = (verb) => {
+const throw_error = (verb) => {
 	throw new Error(
 		ssr
 			? `Can only ${verb} session store in browser`
@@ -76,6 +76,6 @@ export const session = {
 
 		return store.subscribe(fn);
 	},
-	set: () => error('set'),
-	update: () => error('update')
+	set: () => throw_error('set'),
+	update: () => throw_error('update')
 };