+++ packages/kit/src/runtime/server/endpoint.js
@@ -1,5 +1,5 @@
 import { get_single_valued_header } from '../../utils/http.js';
-import { lowercase_keys } from './utils.js';
+import { decode_params, lowercase_keys } from './utils.js';
 
 /** @param {string} body */
 function error(body) {
@@ -50,10 +50,13 @@ export async function render_endpoint(request, route, match) {
 		return;
 	}
 
-	const params = route.params(match);
+	// we're mutating `request` so that we don't have to do { ...request, params }
+	// on the next line, since that breaks the getters that replace path, query and
+	// origin. We could revert that once we remove the getters
+	request.params = route.params ? decode_params(route.params(match)) : {};
 
-	const response = await handler({ ...request, params });
-	const preface = `Invalid response from route ${request.path}`;
+	const response = await handler(request);
+	const preface = `Invalid response from route ${request.url.pathname}`;
 
 	if (!response) {
 		return;