+++ packages/kit/src/runtime/server/page/respond.js
@@ -18,7 +18,7 @@ import { coalesce_to_error } from '../../../utils/error.js';
  *   state: SSRRenderState;
  *   $session: any;
  *   route: import('types/internal').SSRPage;
- *   page: import('types/page').Page;
+ *   params: Record<string, string>;
  * }} opts
  * @returns {Promise<ServerResponse | undefined>}
  */
@@ -29,7 +29,9 @@ export async function respond(opts) {
 	let nodes;
 
 	try {
-		nodes = await Promise.all(route.a.map((id) => (id ? options.load_component(id) : undefined)));
+		nodes = await Promise.all(
+			route.a.map((n) => options.manifest._.nodes[n] && options.manifest._.nodes[n]())
+		);
 	} catch (err) {
 		const error = coalesce_to_error(err);
 
@@ -84,6 +86,7 @@ export async function respond(opts) {
 				try {
 					loaded = await load_node({
 						...opts,
+						url: request.url,
 						node,
 						stuff,
 						prerender_enabled: is_prerender_enabled(options, node, state),
@@ -126,7 +129,7 @@ export async function respond(opts) {
 				if (error) {
 					while (i--) {
 						if (route.b[i]) {
-							const error_node = await options.load_component(route.b[i]);
+							const error_node = await options.manifest._.nodes[route.b[i]]();
 
 							/** @type {Loaded} */
 							let node_loaded;
@@ -140,6 +143,7 @@ export async function respond(opts) {
 								const error_loaded = /** @type {import('./types').Loaded} */ (
 									await load_node({
 										...opts,
+										url: request.url,
 										node: error_node,
 										stuff: node_loaded.stuff,
 										prerender_enabled: is_prerender_enabled(options, error_node, state),
@@ -197,6 +201,7 @@ export async function respond(opts) {
 		return with_cookies(
 			await render_response({
 				...opts,
+				url: request.url,
 				page_config,
 				status,
 				error,