+++ packages/kit/test/apps/basics/src/routes/load/change-detection/one/[x].svelte
@@ -1,14 +1,18 @@
 <script context="module">
+	import { browser } from '$app/env';
+
 	let count = 0;
 
 	/** @type {import('@sveltejs/kit').Load} */
-	export async function load({ page }) {
-		count += 1;
+	export async function load({ params }) {
+		if (browser) {
+			count += 1;
+		}
 
 		return {
 			maxage: 5,
 			props: {
-				x: page.params.x,
+				x: params.x,
 				loads: count
 			}
 		};