+++ packages/kit/test/apps/basics/src/routes/load/server-fetch-request.svelte
@@ -1,9 +1,9 @@
 <script context="module">
 	/** @type {import('@sveltejs/kit').Load} */
-	export async function load({ page, fetch }) {
-		const url = `http://localhost:${page.query.get('port')}/server-fetch-request.json`;
+	export async function load({ url, fetch }) {
+		const href = `http://localhost:${url.searchParams.get('port')}/server-fetch-request.json`;
 
-		const res = await fetch(url);
+		const res = await fetch(href);
 		const { answer } = await res.json();
 		return {
 			props: { answer }