+++ packages/kit/test/apps/basics/svelte.config.js
@@ -1,7 +1,6 @@
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	kit: {
-		hostHeader: 'x-forwarded-host',
 		vite: {
 			build: {
 				minify: false